react-modal: moved modal code to another file
